requirements.txt
Flask
Gunicorn
flask-wtf 
wtforms 
flask-uploads
email_validator
sqlalchemy 
pymysql